requirements.txt
langchain-classic
langchain
langchain-community
langchain-chroma
langchain-ollama
langchain-core
langchain-text-splitters
python-docx
python-dotenv
google-generativeai
ollama
pandas
numpy
tqdm
ipykernel
chromadb
ragas